% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_csquares.R
\name{as_csquares}
\alias{as_csquares}
\alias{as_csquares.character}
\alias{as_csquares.numeric}
\alias{as_csquares.sf}
\alias{as_csquares.stars}
\title{Convert lon-lat coordinates into c-square codes}
\usage{
as_csquares(x, resolution, ...)

\method{as_csquares}{character}(x)

\method{as_csquares}{numeric}(x, resolution = 1)

\method{as_csquares}{sf}(x, resolution = 1)

\method{as_csquares}{stars}(x, resolution = 1)
}
\arguments{
\item{x}{An object to be encoded as c-squares. \code{x} can be a \code{numeric} \code{matrix} with two columns
containing the x and y coordinates. \code{x} can also be a simple features object
(\code{\link[sf:st_as_sf]{sf}}) or a spatial arrays object (\code{\link[stars:st_as_stars]{stars}}).}

\item{resolution}{Resolution (in WGS84 degrees) to be used for creating c-squares codes.
As per c-square specifications, the resolution should be 10 or less, yet greater than 0.
It should be a tenfold of 1 or 5. Valid resolutions are therefore: 10, 5, 1, 0.5, 0.1, etc.}
}
\value{
Returns a \code{vector} with the same number of items as rows in \code{x}. It contains
c-square codes corresponding with \code{x}.
}
\description{
Takes WGS84 longitude and latitude coordinates and finds the closest matching c-squares
for a given resolution.
}
\examples{
as_csquares(cbind(x = 5.2399066, y = 52.7155812), resolution = 1)
}
\author{
Pepijn de Vries
}
