% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquares_methods.R
\name{format.csquares}
\alias{format.csquares}
\alias{print.csquares}
\alias{as.character.csquares}
\alias{summary.csquares}
\alias{as.data.frame.csquares}
\title{Basic csquares methods}
\usage{
\method{format}{csquares}(x, ...)

\method{print}{csquares}(x, ...)

\method{as.character}{csquares}(x, ...)

\method{summary}{csquares}(object, ...)

\method{as.data.frame}{csquares}(x, ...)
}
\arguments{
\item{x, object}{A \code{csquares} object to be handled by the s3 methods}

\item{...}{Ignored}
}
\value{
Returns (a formatted version of) x
}
\description{
Basic S3 methods for \code{csquares} objects for formatting and printing the objects
}
