% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquares_methods.R
\name{format.csquares}
\alias{format.csquares}
\alias{print.csquares}
\alias{as.character.csquares}
\alias{summary.csquares}
\alias{as.data.frame.csquares}
\alias{data.frame.csquares}
\alias{c.csquares}
\alias{rbind.csquares}
\alias{cbind.csquares}
\alias{[.csquares}
\alias{[[.csquares}
\alias{$.csquares}
\alias{[<-.csquares}
\alias{[[<-.csquares}
\alias{$<-.csquares}
\alias{merge.csquares}
\title{Basic csquares methods}
\usage{
\method{format}{csquares}(x, ...)

\method{print}{csquares}(x, ...)

\method{as.character}{csquares}(x, ...)

\method{summary}{csquares}(object, ...)

\method{as.data.frame}{csquares}(x, ...)

data.frame.csquares(...)

\method{c}{csquares}(...)

\method{rbind}{csquares}(..., deparse.level = 1)

\method{cbind}{csquares}(..., deparse.level = 1)

\method{[}{csquares}(x, i, j, ..., drop = FALSE)

\method{[[}{csquares}(x, i)

\method{$}{csquares}(x, name)

\method{[}{csquares}(x, i, j) <- value

\method{[[}{csquares}(x, i) <- value

\method{$}{csquares}(x, i) <- value

\method{merge}{csquares}(x, y, ...)
}
\arguments{
\item{x, object}{A \code{csquares} object to be handled by the s3 methods}

\item{...}{Passed on to generic methods}

\item{deparse.level}{integer controlling the construction of labels in
    the case of non-matrix-like arguments (for the default method):\cr
    \code{deparse.level = 0} constructs no labels;\cr
    the default \code{deparse.level = 1} typically and
    \code{deparse.level = 2} always construct labels from the argument
    names, see the \sQuote{Value} section below.}

\item{i, j, name}{Indices/name for selecting subsets of \code{x}}

\item{drop}{\code{logical} value indicating if unused dimensions should be dropped}

\item{value}{Replacement values for a subset. a \code{csquares} object or a \code{character} string that can be coerced
to a \code{csquares} object}

\item{y}{A \code{data.frame} to be merged with \code{x}}
}
\value{
Returns (a subsetted / formatted / modified version of) x
}
\description{
Basic S3 methods for handling \code{csquares} objects
}
