% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{ices_columns}
\alias{ices_columns}
\title{Valid ICES rectangle columns}
\usage{
ices_columns()
}
\value{
A \code{character} vector with all allowed codes for the columns in ICES
rectangles.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Get all valid column codes of ICES rectangles.
Note that ICES subrectangles are not compatible with csquares. For more details
see \code{vignette("ices")}.
}
\examples{
ices_columns()
}
