% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca.R
\docType{data}
\name{orca}
\alias{orca}
\title{Killer whale realm}
\format{
\subsection{\code{orca}}{

The \code{orca} object is a Killer whale realm data set extracted from the data as provided
by Costello (2017) and published by Costello \emph{et al.} (2017).
It is a data frame with 2,058 rows and two columns:
\describe{
\item{csquares}{c-squares codes indicating spatial grid cells}
\item{orcinus_orca}{\code{logical} values indicating whether the corresponding c-squares
grid cell belongs to the killer whales (Orcinus orca) biogeographic realm or not.}
}
}
}
\usage{
orca
}
\description{
Killer whale realm
}
\references{
\itemize{
\item Costello, M.J. (2017); University of Auckland \doi{10.17608/k6.auckland.5086654}
Licence \href{https://creativecommons.org/licenses/by/4.0/}{CC BY 4.0}
\item Costello M.J., Tsai P., Wong P.S., Cheung A.K.L, Basher Z. & Chaudhary C. (2017);
"Marine biogeographic realms and species endemicity" Nature Communications 8, 1057
\doi{10.1038/s41467-017-01121-2}
}
}
\keyword{datasets}
