% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stars.R
\name{st_as_stars.csquares}
\alias{st_as_stars.csquares}
\title{Coerce csqaures object into a stars object}
\usage{
st_as_stars.csquares(x, ...)
}
\arguments{
\item{x}{An object of class \code{csquares} created with \link{new_csquares} or \link{as_csquares}}

\item{...}{ignored.}
}
\value{
Returns a spatiotemporal array (\link[stars:st_as_stars]{stars}) object based on \code{x}.
}
\description{
Take a \code{csquares} object created with \link{new_csquares} or \link{as_csquares} and
coerce it to a spatiotemporal array (\link[stars:st_as_stars]{stars}).
}
\examples{
library(stars)
st_as_stars(as_csquares("7500:110:3|7500:110:1|1500:110:3|1500:110:1"))
st_as_stars(as_csquares(orca, csquares = "csquares"))
}
\author{
Pepijn de Vries
}
