% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_csquares}
\alias{validate_csquares}
\title{Test if a csquares object is valid}
\usage{
validate_csquares(x)
}
\arguments{
\item{x}{An object of class \code{csquares} to be evaluated.}
}
\value{
Returns a \code{logical} value indicating whether the \code{csquares} object is valid or not.
}
\description{
Tests if a \code{csquares} object is correctly specified and can be translated into valid coordinates
}
\examples{
validate_csquares(
  as_csquares("7500:110:3|7500:110:1|1500:110:3|1500:110:1")
)
}
\author{
Pepijn de Vries
}
