% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{vctrs}
\alias{vctrs}
\alias{vec_cast.csquares}
\alias{vec_cast.csquares.csquares}
\alias{vec_cast.csquares.character}
\alias{vec_cast.csquares.default}
\alias{vec_ptype2.csquares}
\alias{vec_ptype2.csquares.character}
\alias{vec_ptype2.csquares.csquares}
\alias{vec_ptype2.csquares.default}
\title{vctrs methods for csquares objects}
\usage{
vec_cast.csquares(x, to, ...)

\method{vec_cast.csquares}{csquares}(x, to, ...)

\method{vec_cast.csquares}{character}(x, to, ...)

\method{vec_cast.csquares}{default}(x, to, ...)

vec_ptype2.csquares(x, y, ...)

\method{vec_ptype2.csquares}{character}(x, y, ...)

\method{vec_ptype2.csquares}{csquares}(x, y, ...)

\method{vec_ptype2.csquares}{default}(x, y, ..., x_arg = "x", y_arg = "y")
}
\arguments{
\item{x, y}{Vector types.}

\item{to}{Types to cast to. If NULL, \code{x} will be returned as is.}

\item{...}{Ignored.}

\item{x_arg, y_arg}{Argument names for \code{x} and \code{y}.}
}
\description{
Implementations to support csquare vctrs operations. There is no need to call these functions directly.
}
