% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquares_methods.R
\name{format.csquares}
\alias{format.csquares}
\alias{show}
\alias{show.csquares}
\alias{print.csquares}
\title{Basic csquares methods}
\usage{
\method{format}{csquares}(x, ...)

show(x, ...)

\method{show}{csquares}(x, ...)

\method{print}{csquares}(x, ...)
}
\arguments{
\item{x}{A \code{csquares} object to be handled by the s3 methods}

\item{...}{Ignored}
}
\value{
Returns (a formatted version of) x
}
\description{
Basic methods for \code{csquares} objects for formatting and printing the objects
}
